(misc) refactor discovery methods to providers
