+++ generators/client/templates.go
@@ -4,7 +4,7 @@ var templates = map[string]string{
 	"action":        "e3sgR2VuZXJhdGVkV2FybmluZyB9fQoKcGFja2FnZSB7eyAuUGFja2FnZSB9fQoKaW1wb3J0ICgKCSJjb250ZXh0IgoJImVuY29kaW5nL2pzb24iCgkiZm10IgoJInN5bmMiCgkiaW8iCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL3Byb3RvY29sIgoJcnBjY2xpZW50ICJnaXRodWIuY29tL2Nob3JpYS1pby9nby1jaG9yaWEvcHJvdmlkZXJzL2FnZW50L21jb3JwYy9jbGllbnQiCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL3Byb3ZpZGVycy9hZ2VudC9tY29ycGMvcmVwbHlmbXQiCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL3Byb3ZpZGVycy9hZ2VudC9tY29ycGMvZGRsL2FnZW50IgopCgovLyB7eyAuQWN0aW9uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fVJlcXVlc3RlciBwZXJmb3JtcyBhIFJQQyByZXF1ZXN0IHRvIHt7IC5BZ2VudE5hbWUgfCBUb0xvd2VyIH19I3t7IC5BY3Rpb25OYW1lIHwgVG9Mb3dlciB9fQp0eXBlIHt7IC5BY3Rpb25OYW1lIHwgU25ha2VUb0NhbWVsIH19UmVxdWVzdGVyIHN0cnVjdCB7CglyICAgICpyZXF1ZXN0ZXIKCW91dGMgY2hhbiAqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQKfQoKLy8ge3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQgaXMgdGhlIG91dHB1dCBmcm9tIHRoZSB7eyAuQWN0aW9uTmFtZSB8IFRvTG93ZXIgfX0gYWN0aW9uCnR5cGUge3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQgc3RydWN0IHsKCWRldGFpbHMgKlJlc3VsdERldGFpbHMKCXJlcGx5ICAgbWFwW3N0cmluZ11pbnRlcmZhY2V7fQp9CgovLyB7eyAuQWN0aW9uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fVJlc3VsdCBpcyB0aGUgcmVzdWx0IGZyb20gYSB7eyAuQWN0aW9uTmFtZSB8IFRvTG93ZXIgfX0gYWN0aW9uCnR5cGUge3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXN1bHQgc3RydWN0IHsKCWRkbCAqYWdlbnQuRERMCglzdGF0cyAgICpycGNjbGllbnQuU3RhdHMKCW91dHB1dHMgW10qe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQKCXJwY3JlcGxpZXMgW10qcmVwbHlmbXQuUlBDUmVwbHkKCW11IHN5bmMuTXV0ZXgKfQoKZnVuYyAoZCAqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXN1bHQpIFJlbmRlclJlc3VsdHModyBpby5Xcml0ZXIsIGZvcm1hdCBSZW5kZXJGb3JtYXQsIGRpc3BsYXlNb2RlIERpc3BsYXlNb2RlLCB2ZXJib3NlIGJvb2wsIHNpbGVudCBib29sLCBjb2xvcml6ZSBib29sLCBsb2cgTG9nKSBlcnJvciB7CglkLm11LkxvY2soKQoJZGVmZXIgZC5tdS5VbmxvY2soKQoKCWlmIGQuc3RhdHMgPT0gbmlsIHsKCQlyZXR1cm4gZm10LkVycm9yZigicmVzdWx0IHN0YXRzIGlzIG5vdCBzZXQsIHJlc3VsdCB3YXMgbm90IGNvbXBsZXRlZCIpCgl9CgoJcmVzdWx0cyA6PSAmcmVwbHlmbXQuUlBDUmVzdWx0c3sKCQlBZ2VudDogICBkLnN0YXRzLkFnZW50KCksCgkJQWN0aW9uOiAgZC5zdGF0cy5BY3Rpb24oKSwKCQlSZXBsaWVzOiBkLnJwY3JlcGxpZXMsCgkJU3RhdHM6ICAgZC5zdGF0cywKCX0KCglhZGRsLCBlcnIgOj0gZC5kZGwuQWN0aW9uSW50ZXJmYWNlKGQuc3RhdHMuQWN0aW9uKCkpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gZXJyCgl9CgoJc3dpdGNoIGZvcm1hdCB7CgljYXNlIEpTT05Gb3JtYXQ6CgkJcmV0dXJuIHJlc3VsdHMuUmVuZGVySlNPTih3LCBhZGRsKQoJY2FzZSBUYWJsZUZvcm1hdDoKCQlyZXR1cm4gcmVzdWx0cy5SZW5kZXJUYWJsZSh3LCBhZGRsKQoJY2FzZSBUWFRGb290ZXI6CgkJcmVzdWx0cy5SZW5kZXJUWFRGb290ZXIodywgdmVyYm9zZSkKCQlyZXR1cm4gbmlsCglkZWZhdWx0OgoJCXJldHVybiByZXN1bHRzLlJlbmRlclRYVCh3LCBhZGRsLCB2ZXJib3NlLCBzaWxlbnQsIHJlcGx5Zm10LkRpc3BsYXlNb2RlKGRpc3BsYXlNb2RlKSwgY29sb3JpemUsIGxvZykKCX0KfQoKLy8gU3RhdHMgaXMgdGhlIHJwYyByZXF1ZXN0IHN0YXRzCmZ1bmMgKGQgKnt7IC5BY3Rpb25OYW1lIHwgU25ha2VUb0NhbWVsIH19UmVzdWx0KSBTdGF0cygpIFN0YXRzIHsKCXJldHVybiBkLnN0YXRzCn0KCi8vIFJlc3VsdERldGFpbHMgaXMgdGhlIGRldGFpbHMgYWJvdXQgdGhlIHJlcXVlc3QKZnVuYyAoZCAqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQpIFJlc3VsdERldGFpbHMoKSAqUmVzdWx0RGV0YWlscyB7CglyZXR1cm4gZC5kZXRhaWxzCn0KCi8vIEhhc2hNYXAgaXMgdGhlIHJhdyBvdXRwdXQgZGF0YQpmdW5jIChkICp7eyAuQWN0aW9uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fU91dHB1dCkgSGFzaE1hcCgpIG1hcFtzdHJpbmddaW50ZXJmYWNle30gewoJcmV0dXJuIGQucmVwbHkKfQoKLy8gSlNPTiBpcyB0aGUgSlNPTiByZXByZXNlbnRhdGlvbiBvZiB0aGUgb3V0cHV0IGRhdGEKZnVuYyAoZCAqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQpIEpTT04oKSAoW11ieXRlLCBlcnJvcikgewoJcmV0dXJuIGpzb24uTWFyc2hhbChkLnJlcGx5KQp9CgovLyBQYXJzZU91dHB1dCBwYXJzZXMgdGhlIHJlc3VsdCB2YWx1ZSBmcm9tIHRoZSB7eyAuQWN0aW9uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fSBhY3Rpb24gaW50byB0YXJnZXQKZnVuYyAoZCAqe3sgJC5BY3Rpb25OYW1lIHwgU25ha2VUb0NhbWVsIH19T3V0cHV0KSBQYXJzZXt7IC5BY3Rpb25OYW1lIHwgU25ha2VUb0NhbWVsIH19T3V0cHV0KHRhcmdldCBpbnRlcmZhY2V7fSkgZXJyb3IgewoJaiwgZXJyIDo9IGQuSlNPTigpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gZm10LkVycm9yZigiY291bGQgbm90IGFjY2VzcyBwYXlsb2FkOiAlcyIsIGVycikKCX0KCgllcnIgPSBqc29uLlVubWFyc2hhbChqLCB0YXJnZXQpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gZm10LkVycm9yZigiY291bGQgbm90IHVubWFyc2hhbCBKU09OIHBheWxvYWQ6ICVzIiwgZXJyKQoJfQoKCXJldHVybiBuaWwKfQoKLy8gRG8gcGVyZm9ybXMgdGhlIHJlcXVlc3QKZnVuYyAoZCAqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXF1ZXN0ZXIpIERvKGN0eCBjb250ZXh0LkNvbnRleHQpICgqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXN1bHQsIGVycm9yKSB7CglkcmVzIDo9ICZ7eyAuQWN0aW9uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fVJlc3VsdHtkZGw6IGQuci5jbGllbnQuZGRsfQoKCWhhbmRsZXIgOj0gZnVuYyhwciBwcm90b2NvbC5SZXBseSwgciAqcnBjY2xpZW50LlJQQ1JlcGx5KSB7CgkJLy8gZmlsdGVyZWQgYnkgZXhwciBmaWx0ZXIKCQlpZiByID09bmlsIHsKCQkJcmV0dXJuCgkJfQoKCQlvdXRwdXQgOj0gJnt7IC5BY3Rpb25OYW1lIHwgU25ha2VUb0NhbWVsIH19T3V0cHV0ewoJCQlyZXBseTogbWFrZShtYXBbc3RyaW5nXWludGVyZmFjZXt9KSwKCQkJZGV0YWlsczogJlJlc3VsdERldGFpbHN7CgkJCQlzZW5kZXI6ICBwci5TZW5kZXJJRCgpLAoJCQkJY29kZTogICAgaW50KHIuU3RhdHVzY29kZSksCgkJCQltZXNzYWdlOiByLlN0YXR1c21zZywKCQkJCXRzOiAgICAgIHByLlRpbWUoKSwKCQkJfSwKCQl9CgoJCWVyciA6PSBqc29uLlVubWFyc2hhbChyLkRhdGEsICZvdXRwdXQucmVwbHkpCgkJaWYgZXJyICE9IG5pbCB7CgkJCWQuci5jbGllbnQuZXJyb3JmKCJDb3VsZCBub3QgZGVjb2RlIHJlcGx5IGZyb20gJXM6ICVzIiwgcHIuU2VuZGVySUQoKSwgZXJyKQoJCX0KCgkJLy8gY2FsbGVyIHdhbnRzIGEgY2hhbm5lbCBzbyB3ZSBkb250IHJldHVybiBhIHJlc3Vsc2V0IHRvbyAobG90cyBvZiBtZW1vcnkgZXRjKQoJCS8vIHRoaXMgaXMgdW51c2VkIG5vdywgbm8gc3VwcG9ydCBmb3Igc2V0dGluZyBhIGNoYW5uZWwKCQlpZiBkLm91dGMgIT0gbmlsIHsKCQkJZC5vdXRjIDwtIG91dHB1dAoJCQlyZXR1cm4KCQl9CgoJCS8vIGVsc2UgcHJlcGFyZSBvdXIgcmVzdWx0IHNldAoJCWRyZXMubXUuTG9jaygpCgkJZHJlcy5vdXRwdXRzID0gYXBwZW5kKGRyZXMub3V0cHV0cywgb3V0cHV0KQoJCWRyZXMucnBjcmVwbGllcyA9IGFwcGVuZChkcmVzLnJwY3JlcGxpZXMsICZyZXBseWZtdC5SUENSZXBseXsKCQkJU2VuZGVyOiAgIHByLlNlbmRlcklEKCksCgkJCVJQQ1JlcGx5OiByLAoJCX0pCgkJZHJlcy5tdS5VbmxvY2soKQoJfQoKCXJlcywgZXJyIDo9IGQuci5kbyhjdHgsIGhhbmRsZXIpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gbmlsLCBlcnIKCX0KCglkcmVzLnN0YXRzID0gcmVzCgoJcmV0dXJuIGRyZXMsIG5pbAp9CgovLyBFYWNoT3V0cHV0IGl0ZXJhdGVzIG92ZXIgYWxsIHJlc3VsdHMgcmVjZWl2ZWQKZnVuYyAoZCAqe3sgLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXN1bHQpIEVhY2hPdXRwdXQoaCBmdW5jKHIgKnt7IC5BY3Rpb25OYW1lIHwgU25ha2VUb0NhbWVsIH19T3V0cHV0KSkgewoJZm9yIF8sIHJlc3AgOj0gcmFuZ2UgZC5vdXRwdXRzIHsKCQloKHJlc3ApCgl9Cn0KCnt7IHJhbmdlICRuYW1lLCAkaW5wdXQgOj0gLk9wdGlvbmFsSW5wdXRzIH19Ci8vIHt7ICRuYW1lIHwgU25ha2VUb0NhbWVsIH19IGlzIGFuIG9wdGlvbmFsIGlucHV0IHRvIHRoZSB7eyAkLkFjdGlvbk5hbWUgfCBUb0xvd2VyIH19IGFjdGlvbgovLwovLyBEZXNjcmlwdGlvbjoge3sgJGlucHV0LkRlc2NyaXB0aW9uIH19CmZ1bmMgKGQgKnt7ICQuQWN0aW9uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fVJlcXVlc3Rlcikge3sgJG5hbWUgfCBTbmFrZVRvQ2FtZWwgfX0odiB7eyBDaG9yaWFUeXBlVG9Hb1R5cGUgJGlucHV0LlR5cGUgfX0pICp7eyAkLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXF1ZXN0ZXIgewoJZC5yLmFyZ3NbInt7ICRuYW1lIHwgVG9Mb3dlciB9fSJdID0gdgoKCXJldHVybiBkCn0Ke3sgZW5kIH19Cnt7IHJhbmdlICRuYW1lLCAkb3V0cHV0IDo9IC5PdXRwdXRzIH19Cnt7LSAkcmV0dXJuX3R5cGUgOj0gJG91dHB1dC5UeXBlIHwgQ2hvcmlhVHlwZVRvR29UeXBlIC19fQovLyB7eyAkbmFtZSB8IFNuYWtlVG9DYW1lbCB9fSBpcyB0aGUgdmFsdWUgb2YgdGhlIHt7ICRuYW1lIH19IG91dHB1dAovLwovLyBEZXNjcmlwdGlvbjoge3sgJG91dHB1dC5EZXNjcmlwdGlvbiB9fQpmdW5jIChkICp7eyAkLkFjdGlvbk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1PdXRwdXQpIHt7ICRuYW1lIHwgU25ha2VUb0NhbWVsIH19KCkge3sgJHJldHVybl90eXBlIH19IHsKe3stIGlmIGVxICJpbnRlcmZhY2V7fSIgJHJldHVybl90eXBlICAtfX0KCXZhbCwgb2sgOj0gZC5yZXBseVsie3sgJG5hbWUgfX0iXQoJaWYgIW9rIHx8IHZhbCA9PSBuaWwgewoJCS8vIHdlIGhhdmUgdG8gYXZvaWQgcmV0dXJuaW5nIG5pbC4oaW50ZXJmYWNle30pCgkJcmV0dXJuIG5pbAoJfQp7ey0gZWxzZSAtfX0KCXZhbCA6PSBkLnJlcGx5WyJ7eyAkbmFtZSB9fSJdCnt7IGVuZCB9fQoKCXJldHVybiB7eyBDaG9yaWFUeXBlVG9WYWxPZlR5cGUgJG91dHB1dC5UeXBlIH19Cn0Ke3sgZW5kIH19Cg==",
 	"client":        "Ly8gZ2VuZXJhdGVkIGNvZGU7IERPIE5PVCBFRElUCgpwYWNrYWdlIHt7IC5QYWNrYWdlIH19CgppbXBvcnQgKAoJImZtdCIKCSJzeW5jIgoJInRpbWUiCgoJImNvbnRleHQiCgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9jaG9yaWEiCgljb3JlY2xpZW50ICJnaXRodWIuY29tL2Nob3JpYS1pby9nby1jaG9yaWEvY2xpZW50L2NsaWVudCIKCSJnaXRodWIuY29tL2Nob3JpYS1pby9nby1jaG9yaWEvY29uZmlnIgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9zcnZjYWNoZSIKCXJwY2NsaWVudCAiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL3Byb3ZpZGVycy9hZ2VudC9tY29ycGMvY2xpZW50IgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9wcm92aWRlcnMvYWdlbnQvbWNvcnBjL2RkbC9hZ2VudCIKCSJnaXRodWIuY29tL2Nob3JpYS1pby9nby1jaG9yaWEvcHJvdG9jb2wiCgkiZ2l0aHViLmNvbS9zaXJ1cHNlbi9sb2dydXMiCikKCi8vIFN0YXRzIGFyZSB0aGUgc3RhdGlzdGljcyBmb3IgYSByZXF1ZXN0CnR5cGUgU3RhdHMgaW50ZXJmYWNlIHsKCUFnZW50KCkgc3RyaW5nCglBY3Rpb24oKSBzdHJpbmcKCUFsbCgpIGJvb2wKCU5vUmVzcG9uc2VGcm9tKCkgW11zdHJpbmcKCVVuZXhwZWN0ZWRSZXNwb25zZUZyb20oKSBbXXN0cmluZwoJRGlzY292ZXJlZENvdW50KCkgaW50CglEaXNjb3ZlcmVkTm9kZXMoKSAqW11zdHJpbmcKCUZhaWxDb3VudCgpIGludAoJT0tDb3VudCgpIGludAoJUmVzcG9uc2VzQ291bnQoKSBpbnQKCVB1Ymxpc2hEdXJhdGlvbigpICh0aW1lLkR1cmF0aW9uLCBlcnJvcikKCVJlcXVlc3REdXJhdGlvbigpICh0aW1lLkR1cmF0aW9uLCBlcnJvcikKCURpc2NvdmVyeUR1cmF0aW9uKCkgKHRpbWUuRHVyYXRpb24sIGVycm9yKQoJT3ZlcnJpZGVEaXNjb3ZlcnlUaW1lKHN0YXJ0IHRpbWUuVGltZSwgZW5kIHRpbWUuVGltZSkKfQoKLy8gTm9kZVNvdXJjZSBkaXNjb3ZlcnMgbm9kZXMKdHlwZSBOb2RlU291cmNlIGludGVyZmFjZSB7CglSZXNldCgpCglEaXNjb3ZlcihjdHggY29udGV4dC5Db250ZXh0LCBmdyBDaG9yaWFGcmFtZXdvcmssIGZpbHRlcnMgW11GaWx0ZXJGdW5jKSAoW11zdHJpbmcsIGVycm9yKQp9CgovLyBDaG9yaWFGcmFtZXdvcmsgaXMgdGhlIENob3JpYSBmcmFtZXdvcmsKdHlwZSBDaG9yaWFGcmFtZXdvcmsgaW50ZXJmYWNlIHsKCUxvZ2dlcihzdHJpbmcpICpsb2dydXMuRW50cnkKCVNldExvZ2dlcigqbG9ncnVzLkxvZ2dlcikKCUNvbmZpZ3VyYXRpb24oKSAqY29uZmlnLkNvbmZpZwoJTmV3TWVzc2FnZShwYXlsb2FkIHN0cmluZywgYWdlbnQgc3RyaW5nLCBjb2xsZWN0aXZlIHN0cmluZywgbXNnVHlwZSBzdHJpbmcsIHJlcXVlc3QgKmNob3JpYS5NZXNzYWdlKSAobXNnICpjaG9yaWEuTWVzc2FnZSwgZXJyIGVycm9yKQoJTmV3UmVwbHlGcm9tVHJhbnNwb3J0SlNPTihwYXlsb2FkIFtdYnl0ZSwgc2tpcHZhbGlkYXRlIGJvb2wpIChtc2cgcHJvdG9jb2wuUmVwbHksIGVyciBlcnJvcikKCU5ld1RyYW5zcG9ydEZyb21KU09OKGRhdGEgc3RyaW5nKSAobWVzc2FnZSBwcm90b2NvbC5UcmFuc3BvcnRNZXNzYWdlLCBlcnIgZXJyb3IpCglNaWRkbGV3YXJlU2VydmVycygpIChzZXJ2ZXJzIHNydmNhY2hlLlNlcnZlcnMsIGVyciBlcnJvcikKCU5ld0Nvbm5lY3RvcihjdHggY29udGV4dC5Db250ZXh0LCBzZXJ2ZXJzIGZ1bmMoKSAoc3J2Y2FjaGUuU2VydmVycywgZXJyb3IpLCBuYW1lIHN0cmluZywgbG9nZ2VyICpsb2dydXMuRW50cnkpIChjb25uIGNob3JpYS5Db25uZWN0b3IsIGVyciBlcnJvcikKCU5ld1JlcXVlc3RJRCgpIChzdHJpbmcsIGVycm9yKQoJQ2VydG5hbWUoKSBzdHJpbmcKCVBRTFF1ZXJ5Q2VydE5hbWVzKHF1ZXJ5IHN0cmluZykgKFtdc3RyaW5nLCBlcnJvcikKCUNvbG9yaXplKGMgc3RyaW5nLCBmb3JtYXQgc3RyaW5nLCBhIC4uLmludGVyZmFjZXt9KSBzdHJpbmcKfQoKLy8gRmlsdGVyRnVuYyBjYW4gZ2VuZXJhdGUgYSBDaG9yaWEgZmlsdGVyCnR5cGUgRmlsdGVyRnVuYyBmdW5jKGYgKnByb3RvY29sLkZpbHRlcikgZXJyb3IKCi8vIFJlbmRlckZvcm1hdCBpcyB0aGUgZm9ybWF0IHVzZWQgYnkgdGhlIFJlbmRlclJlc3VsdHMgaGVscGVyCnR5cGUgUmVuZGVyRm9ybWF0IGludAoKY29uc3QgKAoJLy8gSlNPTkZvcm1hdCByZW5kZXJzIHRoZSByZXN1bHRzIGFzIGEgSlNPTiBkb2N1bWVudAoJSlNPTkZvcm1hdCBSZW5kZXJGb3JtYXQgPSBpb3RhCgoJLy8gVGV4dEZvcm1hdCByZW5kZXJzIHRoZSByZXN1bHRzIGFzIGEgQ2hvcmlhIHR5cGljYWwgcmVzdWx0IHNldCBpbiBsaW5lIHdpdGggY2hvcmlhIHJlcSBvdXRwdXQKCVRleHRGb3JtYXQKCgkvLyBUYWJsZUZvcm1hdCByZW5kZXJzIGFsbCBzdWNjZXNzZnVsIHJlc3BvbnNlcyBpbiBhIHRhYmxlCglUYWJsZUZvcm1hdAoKCS8vIFRYVEZvb3RlciByZW5kZXJzIG9ubHkgdGhlIHJlcXVlc3Qgc3VtbWFyeSBzdGF0aXN0aWNzCglUWFRGb290ZXIKKQoKLy8gRGlzcGxheU1vZGUgb3ZlcnJpZGVzIHRoZSBEREwgZGlzcGxheSBoaW50cwp0eXBlIERpc3BsYXlNb2RlIHVpbnQ4Cgpjb25zdCAoCgkvLyBEaXNwbGF5RERMIHNob3dzIHJlc3VsdHMgYmFzZWQgb24gdGhlIGNvbmZpZ3VyYXRpb24gaW4gdGhlIERETCBmaWxlCglEaXNwbGF5RERMID0gRGlzcGxheU1vZGUoaW90YSkKCS8vIERpc3BsYXlPSyBzaG93cyBvbmx5IHBhc3NpbmcgcmVzdWx0cwoJRGlzcGxheU9LCgkvLyBEaXNwbGF5RmFpbGVkIHNob3dzIG9ubHkgZmFpbGVkIHJlc3VsdHMKCURpc3BsYXlGYWlsZWQKCS8vIERpc3BsYXlBbGwgc2hvd3MgYWxsIHJlc3VsdHMKCURpc3BsYXlBbGwKCS8vIERpc3BsYXlOb25lIHNob3dzIG5vIHJlc3VsdHMKCURpc3BsYXlOb25lCikKCnR5cGUgTG9nIGludGVyZmFjZSB7CglEZWJ1Z2YoZm9ybWF0IHN0cmluZywgYXJncyAuLi5pbnRlcmZhY2V7fSkKCUluZm9mKGZvcm1hdCBzdHJpbmcsIGFyZ3MgLi4uaW50ZXJmYWNle30pCglXYXJuZihmb3JtYXQgc3RyaW5nLCBhcmdzIC4uLmludGVyZmFjZXt9KQoJRXJyb3JmKGZvcm1hdCBzdHJpbmcsIGFyZ3MgLi4uaW50ZXJmYWNle30pCglGYXRhbGYoZm9ybWF0IHN0cmluZywgYXJncyAuLi5pbnRlcmZhY2V7fSkKCVBhbmljZihmb3JtYXQgc3RyaW5nLCBhcmdzIC4uLmludGVyZmFjZXt9KQp9CgovLyB7eyAuRERMLk1ldGFkYXRhLk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1DbGllbnQgdG8gdGhlIHt7IC5EREwuTWV0YWRhdGEuTmFtZSB9fSBhZ2VudAp0eXBlIHt7IC5EREwuTWV0YWRhdGEuTmFtZSB8IFNuYWtlVG9DYW1lbCB9fUNsaWVudCBzdHJ1Y3QgewoJZncgICAgICAgICAgIENob3JpYUZyYW1ld29yawoJY2ZnICAgICAgICAgICAqY29uZmlnLkNvbmZpZwoJZGRsICAgICAgICAgICAqYWdlbnQuRERMCglucyAgICAgICAgICAgIE5vZGVTb3VyY2UKCWNsaWVudE9wdHMgICAgKmluaXRPcHRpb25zCgljbGllbnRSUENPcHRzIFtdcnBjY2xpZW50LlJlcXVlc3RPcHRpb24KCWZpbHRlcnMgICAgICAgW11GaWx0ZXJGdW5jCgl0YXJnZXRzICAgICAgIFtdc3RyaW5nCgl3b3JrZXJzCSAgICAgIGludAoJZXhwckZpbHRlciAgICBzdHJpbmcKCglzeW5jLk11dGV4Cn0KCi8vIE1ldGFkYXRhIGlzIHRoZSBhZ2VudCBtZXRhZGF0YQp0eXBlIE1ldGFkYXRhIHN0cnVjdCB7CglMaWNlbnNlICAgICBzdHJpbmcgYGpzb246ImxpY2Vuc2UiYAoJQXV0aG9yICAgICAgc3RyaW5nIGBqc29uOiJhdXRob3IiYAoJVGltZW91dCAgICAgaW50ICAgIGBqc29uOiJ0aW1lb3V0ImAKCU5hbWUgICAgICAgIHN0cmluZyBganNvbjoibmFtZSJgCglWZXJzaW9uICAgICBzdHJpbmcgYGpzb246InZlcnNpb24iYAoJVVJMICAgICAgICAgc3RyaW5nIGBqc29uOiJ1cmwiYAoJRGVzY3JpcHRpb24gc3RyaW5nIGBqc29uOiJkZXNjcmlwdGlvbiJgCn0KCi8vIE11c3QgY3JlYXRlIGEgbmV3IGNsaWVudCBhbmQgcGFuaWNzIG9uIGVycm9yCmZ1bmMgTXVzdChvcHRzIC4uLkluaXRpYWxpemF0aW9uT3B0aW9uKSAoY2xpZW50ICp7eyAuRERMLk1ldGFkYXRhLk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1DbGllbnQpIHsKCWMsIGVyciA6PSBOZXcob3B0cy4uLikKCWlmIGVyciAhPSBuaWwgewoJCXBhbmljKGVycikKCX0KCglyZXR1cm4gYwp9CgovLyBOZXcgY3JlYXRlcyBhIG5ldyBjbGllbnQgdG8gdGhlIHt7IC5EREwuTWV0YWRhdGEuTmFtZSB9fSBhZ2VudApmdW5jIE5ldyhvcHRzIC4uLkluaXRpYWxpemF0aW9uT3B0aW9uKSAoY2xpZW50ICp7eyAuRERMLk1ldGFkYXRhLk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1DbGllbnQsIGVyciBlcnJvcikgewoJYyA6PSAme3sgLkRETC5NZXRhZGF0YS5OYW1lIHwgU25ha2VUb0NhbWVsIH19Q2xpZW50ewoJCWRkbDogICAgICAgICAgICZhZ2VudC5EREx7fSwKCQljbGllbnRSUENPcHRzOiBbXXJwY2NsaWVudC5SZXF1ZXN0T3B0aW9ue30sCgkJZmlsdGVyczogICAgICAgW11GaWx0ZXJGdW5jewoJCSAgICBGaWx0ZXJGdW5jKGNvcmVjbGllbnQuQWdlbnRGaWx0ZXIoInt7IC5EREwuTWV0YWRhdGEuTmFtZSB9fSIpKSwKCQl9LAoJCWNsaWVudE9wdHM6ICZpbml0T3B0aW9uc3sKCQkJY2ZnRmlsZTogY2hvcmlhLlVzZXJDb25maWcoKSwKCQl9LAoJCXRhcmdldHM6IFtdc3RyaW5ne30sCgl9CgoJZm9yIF8sIG9wdCA6PSByYW5nZSBvcHRzIHsKCQlvcHQoYy5jbGllbnRPcHRzKQoJfQoKCWMuZncsIGVyciA9IGNob3JpYS5OZXcoYy5jbGllbnRPcHRzLmNmZ0ZpbGUpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gbmlsLCBmbXQuRXJyb3JmKCJjb3VsZCBub3QgaW5pdGlhbGl6ZSBDaG9yaWE6ICVzIiwgZXJyKQoJfQoKCWMuY2ZnID0gYy5mdy5Db25maWd1cmF0aW9uKCkKCglpZiBjLmNsaWVudE9wdHMuZHQgPiAwIHsKCQljLmNmZy5EaXNjb3ZlcnlUaW1lb3V0ID0gaW50KGMuY2xpZW50T3B0cy5kdC5TZWNvbmRzKCkpCgl9CgoJaWYgYy5jbGllbnRPcHRzLm5zID09IG5pbCB7CgkJc3dpdGNoIGMuY2ZnLkRlZmF1bHREaXNjb3ZlcnlNZXRob2QgewoJCWNhc2UgImNob3JpYSI6CgkJCWMuY2xpZW50T3B0cy5ucyA9ICZQdXBwZXREQk5Te30KCQlkZWZhdWx0OgoJCQljLmNsaWVudE9wdHMubnMgPSAmQnJvYWRjYXN0TlN7fQoJCX0KCX0KCWMubnMgPSBjLmNsaWVudE9wdHMubnMKCglpZiBjLmNsaWVudE9wdHMubG9nZ2VyID09IG5pbCB7CgkJYy5jbGllbnRPcHRzLmxvZ2dlciA9IGMuZncuTG9nZ2VyKCJ7eyAuRERMLk1ldGFkYXRhLk5hbWUgfX0iKQoJfSBlbHNlIHsKCQljLmZ3LlNldExvZ2dlcihjLmNsaWVudE9wdHMubG9nZ2VyLkxvZ2dlcikKCX0KCgljLmRkbCwgZXJyID0gRERMKCkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBuaWwsIGZtdC5FcnJvcmYoImNvdWxkIG5vdCBwYXJzZSBlbWJlZGRlZCBEREw6ICVzIiwgZXJyKQoJfQoKCXJldHVybiBjLCBuaWwKfQoKLy8gQWdlbnRNZXRhZGF0YSBpcyB0aGUgYWdlbnQgbWV0YWRhdGEgdGhpcyBjbGllbnQgc3VwcG9ydHMKZnVuYyAocCAqe3sgLkRETC5NZXRhZGF0YS5OYW1lIHwgU25ha2VUb0NhbWVsIH19Q2xpZW50KSBBZ2VudE1ldGFkYXRhKCkgKk1ldGFkYXRhIHsKCXJldHVybiAmTWV0YWRhdGF7CgkJTGljZW5zZTogICAgIHAuZGRsLk1ldGFkYXRhLkxpY2Vuc2UsCgkJQXV0aG9yOiAgICAgIHAuZGRsLk1ldGFkYXRhLkF1dGhvciwKCQlUaW1lb3V0OiAgICAgcC5kZGwuTWV0YWRhdGEuVGltZW91dCwKCQlOYW1lOiAgICAgICAgcC5kZGwuTWV0YWRhdGEuTmFtZSwKCQlWZXJzaW9uOiAgICAgcC5kZGwuTWV0YWRhdGEuVmVyc2lvbiwKCQlVUkw6ICAgICAgICAgcC5kZGwuTWV0YWRhdGEuVVJMLAoJCURlc2NyaXB0aW9uOiBwLmRkbC5NZXRhZGF0YS5EZXNjcmlwdGlvbiwKCX0KfQoKLy8gRGlzY292ZXJOb2RlcyBwZXJmb3JtcyBhIGRpc2NvdmVyeSB1c2luZyB0aGUgY29uZmlndXJlZCBmaWx0ZXIgYW5kIG5vZGUgc291cmNlCmZ1bmMgKHAgKnt7IC5EREwuTWV0YWRhdGEuTmFtZSB8IFNuYWtlVG9DYW1lbCB9fUNsaWVudCkgRGlzY292ZXJOb2RlcyhjdHggY29udGV4dC5Db250ZXh0KSAobm9kZXMgW11zdHJpbmcsIGVyciBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCglyZXR1cm4gcC5ucy5EaXNjb3ZlcihjdHgsIHAuZncsIHAuZmlsdGVycykKfQoKe3sgcmFuZ2UgJGksICRhY3Rpb24gOj0gLkRETC5BY3Rpb25zIH19Ci8vIHt7ICRhY3Rpb24uTmFtZSB8IFNuYWtlVG9DYW1lbCB9fSBwZXJmb3JtcyB0aGUge3sgJGFjdGlvbi5OYW1lIHwgVG9Mb3dlciB9fSBhY3Rpb24KLy8KLy8gRGVzY3JpcHRpb246IHt7ICRhY3Rpb24uRGVzY3JpcHRpb24gfX0Ke3stIGlmIENob3JpYVJlcXVpcmVkSW5wdXRzICRhY3Rpb24gfX0KLy8KLy8gUmVxdWlyZWQgSW5wdXRzOgp7ey0gcmFuZ2UgJG5hbWUsICRpbnB1dCA6PSBDaG9yaWFSZXF1aXJlZElucHV0cyAkYWN0aW9uIH19Ci8vICAgIC0ge3sgJG5hbWUgfX0gKHt7ICRpbnB1dC5UeXBlIHwgQ2hvcmlhVHlwZVRvR29UeXBlIH19KSAtIHt7ICRpbnB1dC5EZXNjcmlwdGlvbiB9fQp7ey0gZW5kIH19Cnt7LSBlbmQgfX0Ke3stIGlmIENob3JpYU9wdGlvbmFsSW5wdXRzICRhY3Rpb24gfX0KLy8KLy8gT3B0aW9uYWwgSW5wdXRzOgp7ey0gcmFuZ2UgJG5hbWUsICRpbnB1dCA6PSBDaG9yaWFPcHRpb25hbElucHV0cyAkYWN0aW9uIH19Ci8vICAgIC0ge3sgJG5hbWUgfX0gKHt7ICRpbnB1dC5UeXBlIHwgQ2hvcmlhVHlwZVRvR29UeXBlIH19KSAtIHt7ICRpbnB1dC5EZXNjcmlwdGlvbiB9fQp7ey0gZW5kIH19Cnt7LSBlbmQgfX0KZnVuYyAocCAqe3sgJC5EREwuTWV0YWRhdGEuTmFtZSB8IFNuYWtlVG9DYW1lbCB9fUNsaWVudCkge3sgJGFjdGlvbi5OYW1lIHwgU25ha2VUb0NhbWVsIH19KHt7ICRhY3Rpb24gfCBDaG9yaWFSZXF1aXJlZElucHV0c1RvRnVuY0FyZ3MgfX0pICp7eyAkYWN0aW9uLk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1SZXF1ZXN0ZXIgewoJZCA6PSAme3sgJGFjdGlvbi5OYW1lIHwgU25ha2VUb0NhbWVsIH19UmVxdWVzdGVyewoJCW91dGM6IG5pbCwKCQlyOiAmcmVxdWVzdGVyewoJCQlhcmdzOiAgIG1hcFtzdHJpbmddaW50ZXJmYWNle317Cnt7LSByYW5nZSAkbmFtZSwgJGlucHV0IDo9IENob3JpYVJlcXVpcmVkSW5wdXRzICRhY3Rpb24gfX0KCQkJCSJ7eyAkbmFtZSB9fSI6IHt7ICRuYW1lIH19SSwKe3stIGVuZCB9fQoJCQl9LAoJCQlhY3Rpb246ICJ7eyAkYWN0aW9uLk5hbWUgfCBUb0xvd2VyIH19IiwKCQkJY2xpZW50OiBwLAoJCX0sCgl9CgoJcmV0dXJuIGQKfQp7eyBlbmQgfX0K",
 	"ddl":           "Ly8gZ2VuZXJhdGVkIGNvZGU7IERPIE5PVCBFRElUCgpwYWNrYWdlIHt7IC5QYWNrYWdlIH19CgppbXBvcnQgKAoJImVuY29kaW5nL2Jhc2U2NCIKCSJlbmNvZGluZy9qc29uIgoKCSJnaXRodWIuY29tL2Nob3JpYS1pby9nby1jaG9yaWEvcHJvdmlkZXJzL2FnZW50L21jb3JwYy9kZGwvYWdlbnQiCikKCgp2YXIgcmF3RERMID0gInt7IC5SYXdEREwgfCBCYXNlNjRFbmNvZGUgfX0iCgovLyBERExCeXRlcyBpcyB0aGUgcmF3IEpTT04gZW5jb2RlZCBEREwgZmlsZSBmb3IgdGhlIGFnZW50CmZ1bmMgRERMQnl0ZXMoKSAoW11ieXRlLCBlcnJvcikgewoJcmV0dXJuIGJhc2U2NC5TdGRFbmNvZGluZy5EZWNvZGVTdHJpbmcocmF3RERMKQp9CgovLyBEREwgaXMgYSBwYXJzZWQgYW5kIGxvYWRlZCBEREwgZm9yIHRoZSBhZ2VudApmdW5jIERETCgpICgqYWdlbnQuRERMLCBlcnJvcikgewoJZGRsaiwgZXJyIDo9IERETEJ5dGVzKCkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBuaWwsIGVycgoJfQoKCWRkbCA6PSAmYWdlbnQuRERMe30KCWVyciA9IGpzb24uVW5tYXJzaGFsKGRkbGosIGRkbCkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBuaWwsIGVycgoJfQoKCXJldHVybiBkZGwsIG5pbAp9Cg==",
-	"discover":      "Ly8gZ2VuZXJhdGVkIGNvZGU7IERPIE5PVCBFRElUCgpwYWNrYWdlIHt7IC5QYWNrYWdlIH19CgppbXBvcnQgKAoJImNvbnRleHQiCgkiZm10IgoJInN5bmMiCgkidGltZSIKCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL2NsaWVudC9kaXNjb3ZlcnkiCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL2NsaWVudC9kaXNjb3ZlcnkvYnJvYWRjYXN0IgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9jbGllbnQvZGlzY292ZXJ5L2V4dGVybmFsIgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9jbGllbnQvZGlzY292ZXJ5L3B1cHBldGRiIgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9wcm90b2NvbCIKKQoKLy8gQnJvYWRjYXN0TlMgaXMgYSBOb2RlU291cmNlIHRoYXQgdXNlcyB0aGUgQ2hvcmlhIG5ldHdvcmsgYnJvYWRjYXN0IG1ldGhvZCB0byBkaXNjb3ZlciBub2Rlcwp0eXBlIEJyb2FkY2FzdE5TIHN0cnVjdCB7Cglub2RlQ2FjaGUgW11zdHJpbmcKCWYgICAgICAgICAqcHJvdG9jb2wuRmlsdGVyCgoJc3luYy5NdXRleAp9CgovLyBSZXNldCByZXNldHMgdGhlIGludGVybmFsIG5vZGUgY2FjaGUKZnVuYyAoYiAqQnJvYWRjYXN0TlMpIFJlc2V0KCkgewoJYi5Mb2NrKCkKCWRlZmVyIGIuVW5sb2NrKCkKCgliLm5vZGVDYWNoZSA9IFtdc3RyaW5ne30KfQoKLy8gRGlzY292ZXIgcGVyZm9ybXMgdGhlIGRpc2NvdmVyeSBvZiBub2RlcyBhZ2FpbnN0IHRoZSBDaG9yaWEgTmV0d29yawpmdW5jIChiICpCcm9hZGNhc3ROUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBmaWx0ZXJzIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJYi5Mb2NrKCkKCWRlZmVyIGIuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKGIubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgYi5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKGIubm9kZUNhY2hlID09IG5pbCB8fCBsZW4oYi5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCXZhciBlcnIgZXJyb3IKCgliLmYsIGVyciA9IHBhcnNlRmlsdGVycyhmaWx0ZXJzKQoJaWYgZXJyICE9IG5pbCB7CgkJcmV0dXJuIG5pbCwgZXJyCgl9CgoJaWYgYi5ub2RlQ2FjaGUgPT0gbmlsIHsKCQliLm5vZGVDYWNoZSA9IFtdc3RyaW5ne30KCX0KCgljZmcgOj0gZncuQ29uZmlndXJhdGlvbigpCglub2RlcywgZXJyIDo9IGJyb2FkY2FzdC5OZXcoZncpLkRpc2NvdmVyKGN0eCwgYnJvYWRjYXN0LkZpbHRlcihiLmYpLCBicm9hZGNhc3QuVGltZW91dCh0aW1lLlNlY29uZCp0aW1lLkR1cmF0aW9uKGNmZy5EaXNjb3ZlcnlUaW1lb3V0KSkpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gW11zdHJpbmd7fSwgZXJyCgl9CgoJYi5ub2RlQ2FjaGUgPSBub2RlcwoKCXJldHVybiBjb3BpZXIoKSwgbmlsCn0KCi8vIEV4dGVybmFsTlMgaXMgYSBOb2RlU291cmNlIHRoYXQgY2FsbHMgYW4gZXh0ZXJuYWwgY29tbWFuZCBmb3IgZGlzY292ZXJ5CnR5cGUgRXh0ZXJuYWxOUyBzdHJ1Y3QgewoJbm9kZUNhY2hlIFtdc3RyaW5nCglmICAgICAgICAgKnByb3RvY29sLkZpbHRlcgoKCXN5bmMuTXV0ZXgKfQoKLy8gUmVzZXQgcmVzZXRzIHRoZSBpbnRlcm5hbCBub2RlIGNhY2hlCmZ1bmMgKHAgKkV4dGVybmFsTlMpIFJlc2V0KCkgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCglwLm5vZGVDYWNoZSA9IFtdc3RyaW5ne30KfQoKZnVuYyAocCAqRXh0ZXJuYWxOUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBmaWx0ZXJzIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKHAubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgcC5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKHAubm9kZUNhY2hlID09IG5pbCB8fCBsZW4ocC5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCXZhciBlcnIgZXJyb3IKCXAuZiwgZXJyID0gcGFyc2VGaWx0ZXJzKGZpbHRlcnMpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gbmlsLCBlcnIKCX0KCglpZiBwLm5vZGVDYWNoZSA9PSBuaWwgewoJCXAubm9kZUNhY2hlID0gW11zdHJpbmd7fQoJfQoKCWNmZyA6PSBmdy5Db25maWd1cmF0aW9uKCkKCW5vZGVzLCBlcnIgOj0gZXh0ZXJuYWwuTmV3KGZ3KS5EaXNjb3ZlcihjdHgsIGV4dGVybmFsLkZpbHRlcihwLmYpLCBleHRlcm5hbC5UaW1lb3V0KHRpbWUuU2Vjb25kKnRpbWUuRHVyYXRpb24oY2ZnLkRpc2NvdmVyeVRpbWVvdXQpKSkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBbXXN0cmluZ3t9LCBlcnIKCX0KCglwLm5vZGVDYWNoZSA9IG5vZGVzCgoJcmV0dXJuIGNvcGllcigpLCBuaWwKfQoKLy8gUHVwcGV0REJOUyBpcyBhIE5vZGVTb3VyY2UgdGhhdCB1c2VzIHRoZSBQdXBwZXREQiBQUUwgUXVlcmllcyB0byBkaXNjb3ZlciBub2Rlcwp0eXBlIFB1cHBldERCTlMgc3RydWN0IHsKCW5vZGVDYWNoZSBbXXN0cmluZwoJZiAgICAgICAgICpwcm90b2NvbC5GaWx0ZXIKCglzeW5jLk11dGV4Cn0KCi8vIFJlc2V0IHJlc2V0cyB0aGUgaW50ZXJuYWwgbm9kZSBjYWNoZQpmdW5jIChwICpQdXBwZXREQk5TKSBSZXNldCgpIHsKCXAuTG9jaygpCglkZWZlciBwLlVubG9jaygpCgoJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cn0KCi8vIERpc2NvdmVyIHBlcmZvcm1zIHRoZSBkaXNjb3Zlcnkgb2Ygbm9kZXMgYWdhaW5zdCB0aGUgQ2hvcmlhIE5ldHdvcmsKZnVuYyAocCAqUHVwcGV0REJOUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBmaWx0ZXJzIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKHAubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgcC5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKHAubm9kZUNhY2hlID09IG5pbCB8fCBsZW4ocC5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCXZhciBlcnIgZXJyb3IKCXAuZiwgZXJyID0gcGFyc2VGaWx0ZXJzKGZpbHRlcnMpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gbmlsLCBlcnIKCX0KCglpZiBsZW4ocC5mLkNvbXBvdW5kKSA+IDAgewoJCXJldHVybiBuaWwsIGZtdC5FcnJvcmYoImNvbXBvdW5kIGZpbHRlcnMgYXJlIG5vdCBzdXBwb3J0ZWQgYnkgUHVwcGV0REIiKQoJfQoKCWlmIHAubm9kZUNhY2hlID09IG5pbCB7CgkJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cgl9CgoJY2ZnIDo9IGZ3LkNvbmZpZ3VyYXRpb24oKQoJbm9kZXMsIGVyciA6PSBwdXBwZXRkYi5OZXcoZncpLkRpc2NvdmVyKGN0eCwgcHVwcGV0ZGIuRmlsdGVyKHAuZiksIHB1cHBldGRiLlRpbWVvdXQodGltZS5TZWNvbmQqdGltZS5EdXJhdGlvbihjZmcuRGlzY292ZXJ5VGltZW91dCkpKQoJaWYgZXJyICE9IG5pbCB7CgkJcmV0dXJuIFtdc3RyaW5ne30sIGVycgoJfQoKCXAubm9kZUNhY2hlID0gbm9kZXMKCglyZXR1cm4gY29waWVyKCksIG5pbAp9CgovLyBNZXRhTlMgaXMgYSBOb2RlU291cmNlIHRoYXQgYXNzaXN0cyBDTEkgdG9vbHMgaW4gY3JlYXRpbmcgQ2hvcmlhIHN0YW5kYXJkIGNvbW1hbmQgbGluZSBiYXNlZCBkaXNjb3ZlcnkuCnR5cGUgTWV0YU5TIHN0cnVjdCB7CgkvLyBPcHRpb25zIGlzIHRoZSBDTEkgb3B0aW9ucyB0byBkaXNjb3ZlciBiYXNlZCBvbgoJT3B0aW9ucyAqZGlzY292ZXJ5LlN0YW5kYXJkT3B0aW9ucwoKCS8vIEFnZW50IHNob3VsZCBiZSB0aGUgYWdlbnQgdGhlIHJlcXVlc3QgaXMgdGFyZ2V0ZWQgYXQKCUFnZW50IHN0cmluZwoKCS8vIERpc2FibGVQaXBlZERpc2NvdmVyeSBwcmV2ZW50cyB0aGUgU1RESU4gYmVpbmcgdXNlZCBhcyBhIGRpc2NvdmVyeSBzb3VyY2UKCURpc2FibGVQaXBlZERpc2NvdmVyeSBib29sCgoJbm9kZUNhY2hlIFtdc3RyaW5nCglzeW5jLk11dGV4Cn0KCi8vIE5ld01ldGFOUyBjcmVhdGVzIGEgbmV3IG1ldGEgZGlzY292ZXJ5IG5vZGUgc291cmNlCmZ1bmMgTmV3TWV0YU5TKG9wdHMgKmRpc2NvdmVyeS5TdGFuZGFyZE9wdGlvbnMsIGVuYWJsZVBpcGVNb2RlIGJvb2wpICpNZXRhTlMgewoJcmV0dXJuICZNZXRhTlN7CgkJT3B0aW9uczogICAgICAgICAgICAgICBvcHRzLAoJCUFnZW50OiAgICAgICAgICAgICAgICAgInt7IC5EREwuTWV0YWRhdGEuTmFtZSB9fSIsCgkJRGlzYWJsZVBpcGVkRGlzY292ZXJ5OiAhZW5hYmxlUGlwZU1vZGUsCgkJbm9kZUNhY2hlOiAgICAgICAgICAgICBbXXN0cmluZ3t9LAoJfQp9CgovLyBSZXNldCByZXNldHMgdGhlIGludGVybmFsIG5vZGUgY2FjaGUKZnVuYyAocCAqTWV0YU5TKSBSZXNldCgpIHsKCXAuTG9jaygpCglkZWZlciBwLlVubG9jaygpCgoJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cn0KCi8vIERpc2NvdmVyIHBlcmZvcm1zIHRoZSBkaXNjb3Zlcnkgb2Ygbm9kZXMgYWdhaW5zdCB0aGUgQ2hvcmlhIE5ldHdvcmsuCmZ1bmMgKHAgKk1ldGFOUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBfIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKHAubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgcC5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKHAubm9kZUNhY2hlID09IG5pbCB8fCBsZW4ocC5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCWlmIHAubm9kZUNhY2hlID09IG5pbCB7CgkJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cgl9CgoJaWYgcC5PcHRpb25zID09IG5pbCB7CgkJcmV0dXJuIG5pbCwgZm10LkVycm9yZigib3B0aW9ucyBoYXZlIG5vdCBiZWVuIHNldCIpCgl9CgoJbm9kZXMsIF8sIGVyciA6PSBwLk9wdGlvbnMuRGlzY292ZXIoY3R4LCBmdywgcC5BZ2VudCwgIXAuRGlzYWJsZVBpcGVkRGlzY292ZXJ5LCBmYWxzZSwgZncuTG9nZ2VyKCJkaXNjb3ZlcnkiKSkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBuaWwsIGVycgoJfQoKCXAubm9kZUNhY2hlID0gbm9kZXMKCglyZXR1cm4gY29waWVyKCksIG5pbAp9CgpmdW5jIHBhcnNlRmlsdGVycyhmcyBbXUZpbHRlckZ1bmMpICgqcHJvdG9jb2wuRmlsdGVyLCBlcnJvcikgewoJZmlsdGVyIDo9IHByb3RvY29sLk5ld0ZpbHRlcigpCgoJZm9yIF8sIGYgOj0gcmFuZ2UgZnMgewoJCWVyciA6PSBmKGZpbHRlcikKCQlpZiBlcnIgIT0gbmlsIHsKCQkJcmV0dXJuIG5pbCwgZXJyCgkJfQoJfQoKCXJldHVybiBmaWx0ZXIsIG5pbAp9Cg==",
+	"discover":      "Ly8gZ2VuZXJhdGVkIGNvZGU7IERPIE5PVCBFRElUCgpwYWNrYWdlIHt7IC5QYWNrYWdlIH19CgppbXBvcnQgKAoJImNvbnRleHQiCgkiZm10IgoJInN5bmMiCgkidGltZSIKCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL2NsaWVudC9kaXNjb3ZlcnkiCgkiZ2l0aHViLmNvbS9jaG9yaWEtaW8vZ28tY2hvcmlhL3Byb3ZpZGVycy9kaXNjb3ZlcnkvYnJvYWRjYXN0IgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9wcm92aWRlcnMvZGlzY292ZXJ5L2V4dGVybmFsIgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9wcm92aWRlcnMvZGlzY292ZXJ5L3B1cHBldGRiIgoJImdpdGh1Yi5jb20vY2hvcmlhLWlvL2dvLWNob3JpYS9wcm90b2NvbCIKKQoKLy8gQnJvYWRjYXN0TlMgaXMgYSBOb2RlU291cmNlIHRoYXQgdXNlcyB0aGUgQ2hvcmlhIG5ldHdvcmsgYnJvYWRjYXN0IG1ldGhvZCB0byBkaXNjb3ZlciBub2Rlcwp0eXBlIEJyb2FkY2FzdE5TIHN0cnVjdCB7Cglub2RlQ2FjaGUgW11zdHJpbmcKCWYgICAgICAgICAqcHJvdG9jb2wuRmlsdGVyCgoJc3luYy5NdXRleAp9CgovLyBSZXNldCByZXNldHMgdGhlIGludGVybmFsIG5vZGUgY2FjaGUKZnVuYyAoYiAqQnJvYWRjYXN0TlMpIFJlc2V0KCkgewoJYi5Mb2NrKCkKCWRlZmVyIGIuVW5sb2NrKCkKCgliLm5vZGVDYWNoZSA9IFtdc3RyaW5ne30KfQoKLy8gRGlzY292ZXIgcGVyZm9ybXMgdGhlIGRpc2NvdmVyeSBvZiBub2RlcyBhZ2FpbnN0IHRoZSBDaG9yaWEgTmV0d29yawpmdW5jIChiICpCcm9hZGNhc3ROUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBmaWx0ZXJzIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJYi5Mb2NrKCkKCWRlZmVyIGIuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKGIubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgYi5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKGIubm9kZUNhY2hlID09IG5pbCB8fCBsZW4oYi5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCXZhciBlcnIgZXJyb3IKCgliLmYsIGVyciA9IHBhcnNlRmlsdGVycyhmaWx0ZXJzKQoJaWYgZXJyICE9IG5pbCB7CgkJcmV0dXJuIG5pbCwgZXJyCgl9CgoJaWYgYi5ub2RlQ2FjaGUgPT0gbmlsIHsKCQliLm5vZGVDYWNoZSA9IFtdc3RyaW5ne30KCX0KCgljZmcgOj0gZncuQ29uZmlndXJhdGlvbigpCglub2RlcywgZXJyIDo9IGJyb2FkY2FzdC5OZXcoZncpLkRpc2NvdmVyKGN0eCwgYnJvYWRjYXN0LkZpbHRlcihiLmYpLCBicm9hZGNhc3QuVGltZW91dCh0aW1lLlNlY29uZCp0aW1lLkR1cmF0aW9uKGNmZy5EaXNjb3ZlcnlUaW1lb3V0KSkpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gW11zdHJpbmd7fSwgZXJyCgl9CgoJYi5ub2RlQ2FjaGUgPSBub2RlcwoKCXJldHVybiBjb3BpZXIoKSwgbmlsCn0KCi8vIEV4dGVybmFsTlMgaXMgYSBOb2RlU291cmNlIHRoYXQgY2FsbHMgYW4gZXh0ZXJuYWwgY29tbWFuZCBmb3IgZGlzY292ZXJ5CnR5cGUgRXh0ZXJuYWxOUyBzdHJ1Y3QgewoJbm9kZUNhY2hlIFtdc3RyaW5nCglmICAgICAgICAgKnByb3RvY29sLkZpbHRlcgoKCXN5bmMuTXV0ZXgKfQoKLy8gUmVzZXQgcmVzZXRzIHRoZSBpbnRlcm5hbCBub2RlIGNhY2hlCmZ1bmMgKHAgKkV4dGVybmFsTlMpIFJlc2V0KCkgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCglwLm5vZGVDYWNoZSA9IFtdc3RyaW5ne30KfQoKZnVuYyAocCAqRXh0ZXJuYWxOUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBmaWx0ZXJzIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKHAubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgcC5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKHAubm9kZUNhY2hlID09IG5pbCB8fCBsZW4ocC5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCXZhciBlcnIgZXJyb3IKCXAuZiwgZXJyID0gcGFyc2VGaWx0ZXJzKGZpbHRlcnMpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gbmlsLCBlcnIKCX0KCglpZiBwLm5vZGVDYWNoZSA9PSBuaWwgewoJCXAubm9kZUNhY2hlID0gW11zdHJpbmd7fQoJfQoKCWNmZyA6PSBmdy5Db25maWd1cmF0aW9uKCkKCW5vZGVzLCBlcnIgOj0gZXh0ZXJuYWwuTmV3KGZ3KS5EaXNjb3ZlcihjdHgsIGV4dGVybmFsLkZpbHRlcihwLmYpLCBleHRlcm5hbC5UaW1lb3V0KHRpbWUuU2Vjb25kKnRpbWUuRHVyYXRpb24oY2ZnLkRpc2NvdmVyeVRpbWVvdXQpKSkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBbXXN0cmluZ3t9LCBlcnIKCX0KCglwLm5vZGVDYWNoZSA9IG5vZGVzCgoJcmV0dXJuIGNvcGllcigpLCBuaWwKfQoKLy8gUHVwcGV0REJOUyBpcyBhIE5vZGVTb3VyY2UgdGhhdCB1c2VzIHRoZSBQdXBwZXREQiBQUUwgUXVlcmllcyB0byBkaXNjb3ZlciBub2Rlcwp0eXBlIFB1cHBldERCTlMgc3RydWN0IHsKCW5vZGVDYWNoZSBbXXN0cmluZwoJZiAgICAgICAgICpwcm90b2NvbC5GaWx0ZXIKCglzeW5jLk11dGV4Cn0KCi8vIFJlc2V0IHJlc2V0cyB0aGUgaW50ZXJuYWwgbm9kZSBjYWNoZQpmdW5jIChwICpQdXBwZXREQk5TKSBSZXNldCgpIHsKCXAuTG9jaygpCglkZWZlciBwLlVubG9jaygpCgoJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cn0KCi8vIERpc2NvdmVyIHBlcmZvcm1zIHRoZSBkaXNjb3Zlcnkgb2Ygbm9kZXMgYWdhaW5zdCB0aGUgQ2hvcmlhIE5ldHdvcmsKZnVuYyAocCAqUHVwcGV0REJOUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBmaWx0ZXJzIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKHAubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgcC5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKHAubm9kZUNhY2hlID09IG5pbCB8fCBsZW4ocC5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCXZhciBlcnIgZXJyb3IKCXAuZiwgZXJyID0gcGFyc2VGaWx0ZXJzKGZpbHRlcnMpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gbmlsLCBlcnIKCX0KCglpZiBsZW4ocC5mLkNvbXBvdW5kKSA+IDAgewoJCXJldHVybiBuaWwsIGZtdC5FcnJvcmYoImNvbXBvdW5kIGZpbHRlcnMgYXJlIG5vdCBzdXBwb3J0ZWQgYnkgUHVwcGV0REIiKQoJfQoKCWlmIHAubm9kZUNhY2hlID09IG5pbCB7CgkJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cgl9CgoJY2ZnIDo9IGZ3LkNvbmZpZ3VyYXRpb24oKQoJbm9kZXMsIGVyciA6PSBwdXBwZXRkYi5OZXcoZncpLkRpc2NvdmVyKGN0eCwgcHVwcGV0ZGIuRmlsdGVyKHAuZiksIHB1cHBldGRiLlRpbWVvdXQodGltZS5TZWNvbmQqdGltZS5EdXJhdGlvbihjZmcuRGlzY292ZXJ5VGltZW91dCkpKQoJaWYgZXJyICE9IG5pbCB7CgkJcmV0dXJuIFtdc3RyaW5ne30sIGVycgoJfQoKCXAubm9kZUNhY2hlID0gbm9kZXMKCglyZXR1cm4gY29waWVyKCksIG5pbAp9CgovLyBNZXRhTlMgaXMgYSBOb2RlU291cmNlIHRoYXQgYXNzaXN0cyBDTEkgdG9vbHMgaW4gY3JlYXRpbmcgQ2hvcmlhIHN0YW5kYXJkIGNvbW1hbmQgbGluZSBiYXNlZCBkaXNjb3ZlcnkuCnR5cGUgTWV0YU5TIHN0cnVjdCB7CgkvLyBPcHRpb25zIGlzIHRoZSBDTEkgb3B0aW9ucyB0byBkaXNjb3ZlciBiYXNlZCBvbgoJT3B0aW9ucyAqZGlzY292ZXJ5LlN0YW5kYXJkT3B0aW9ucwoKCS8vIEFnZW50IHNob3VsZCBiZSB0aGUgYWdlbnQgdGhlIHJlcXVlc3QgaXMgdGFyZ2V0ZWQgYXQKCUFnZW50IHN0cmluZwoKCS8vIERpc2FibGVQaXBlZERpc2NvdmVyeSBwcmV2ZW50cyB0aGUgU1RESU4gYmVpbmcgdXNlZCBhcyBhIGRpc2NvdmVyeSBzb3VyY2UKCURpc2FibGVQaXBlZERpc2NvdmVyeSBib29sCgoJbm9kZUNhY2hlIFtdc3RyaW5nCglzeW5jLk11dGV4Cn0KCi8vIE5ld01ldGFOUyBjcmVhdGVzIGEgbmV3IG1ldGEgZGlzY292ZXJ5IG5vZGUgc291cmNlCmZ1bmMgTmV3TWV0YU5TKG9wdHMgKmRpc2NvdmVyeS5TdGFuZGFyZE9wdGlvbnMsIGVuYWJsZVBpcGVNb2RlIGJvb2wpICpNZXRhTlMgewoJcmV0dXJuICZNZXRhTlN7CgkJT3B0aW9uczogICAgICAgICAgICAgICBvcHRzLAoJCUFnZW50OiAgICAgICAgICAgICAgICAgInt7IC5EREwuTWV0YWRhdGEuTmFtZSB9fSIsCgkJRGlzYWJsZVBpcGVkRGlzY292ZXJ5OiAhZW5hYmxlUGlwZU1vZGUsCgkJbm9kZUNhY2hlOiAgICAgICAgICAgICBbXXN0cmluZ3t9LAoJfQp9CgovLyBSZXNldCByZXNldHMgdGhlIGludGVybmFsIG5vZGUgY2FjaGUKZnVuYyAocCAqTWV0YU5TKSBSZXNldCgpIHsKCXAuTG9jaygpCglkZWZlciBwLlVubG9jaygpCgoJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cn0KCi8vIERpc2NvdmVyIHBlcmZvcm1zIHRoZSBkaXNjb3Zlcnkgb2Ygbm9kZXMgYWdhaW5zdCB0aGUgQ2hvcmlhIE5ldHdvcmsuCmZ1bmMgKHAgKk1ldGFOUykgRGlzY292ZXIoY3R4IGNvbnRleHQuQ29udGV4dCwgZncgQ2hvcmlhRnJhbWV3b3JrLCBfIFtdRmlsdGVyRnVuYykgKFtdc3RyaW5nLCBlcnJvcikgewoJcC5Mb2NrKCkKCWRlZmVyIHAuVW5sb2NrKCkKCgljb3BpZXIgOj0gZnVuYygpIFtdc3RyaW5nIHsKCQlvdXQgOj0gbWFrZShbXXN0cmluZywgbGVuKHAubm9kZUNhY2hlKSkKCQljb3B5KG91dCwgcC5ub2RlQ2FjaGUpCgoJCXJldHVybiBvdXQKCX0KCglpZiAhKHAubm9kZUNhY2hlID09IG5pbCB8fCBsZW4ocC5ub2RlQ2FjaGUpID09IDApIHsKCQlyZXR1cm4gY29waWVyKCksIG5pbAoJfQoKCWlmIHAubm9kZUNhY2hlID09IG5pbCB7CgkJcC5ub2RlQ2FjaGUgPSBbXXN0cmluZ3t9Cgl9CgoJaWYgcC5PcHRpb25zID09IG5pbCB7CgkJcmV0dXJuIG5pbCwgZm10LkVycm9yZigib3B0aW9ucyBoYXZlIG5vdCBiZWVuIHNldCIpCgl9CgoJbm9kZXMsIF8sIGVyciA6PSBwLk9wdGlvbnMuRGlzY292ZXIoY3R4LCBmdywgcC5BZ2VudCwgIXAuRGlzYWJsZVBpcGVkRGlzY292ZXJ5LCBmYWxzZSwgZncuTG9nZ2VyKCJkaXNjb3ZlcnkiKSkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybiBuaWwsIGVycgoJfQoKCXAubm9kZUNhY2hlID0gbm9kZXMKCglyZXR1cm4gY29waWVyKCksIG5pbAp9CgpmdW5jIHBhcnNlRmlsdGVycyhmcyBbXUZpbHRlckZ1bmMpICgqcHJvdG9jb2wuRmlsdGVyLCBlcnJvcikgewoJZmlsdGVyIDo9IHByb3RvY29sLk5ld0ZpbHRlcigpCgoJZm9yIF8sIGYgOj0gcmFuZ2UgZnMgewoJCWVyciA6PSBmKGZpbHRlcikKCQlpZiBlcnIgIT0gbmlsIHsKCQkJcmV0dXJuIG5pbCwgZXJyCgkJfQoJfQoKCXJldHVybiBmaWx0ZXIsIG5pbAp9Cg==",
 	"doc":           "e3sgR2VuZXJhdGVkV2FybmluZyB9fQoKLy8gUGFja2FnZSB7eyAuUGFja2FnZSB9fSBpcyBhbiBBUEkgY2xpZW50IHRvIHRoZSBDaG9yaWEge3sgLkRETC5NZXRhZGF0YS5OYW1lIHwgQ2FwaXRhbGl6ZSB9fSBhZ2VudCBWZXJzaW9uIHt7IC5EREwuTWV0YWRhdGEuVmVyc2lvbiB9fS4KLy8KLy8gQWN0aW9uczoKe3stIHJhbmdlICRpLCAkYWN0aW9uIDo9IC5EREwuQWN0aW9ucyB9fQovLyAgICoge3sgJGFjdGlvbi5OYW1lIHwgU25ha2VUb0NhbWVsIH19IC0ge3sgJGFjdGlvbi5EZXNjcmlwdGlvbiAtfX0Ke3sgZW5kIH19CnBhY2thZ2Uge3sgLlBhY2thZ2UgfX0KCg==",
 	"initoptions":   "Ly8gZ2VuZXJhdGVkIGNvZGU7IERPIE5PVCBFRElUCgpwYWNrYWdlIHt7IC5QYWNrYWdlIH19CgppbXBvcnQgKAoJInRpbWUiCgkiZ2l0aHViLmNvbS9zaXJ1cHNlbi9sb2dydXMiCikKCnR5cGUgaW5pdE9wdGlvbnMgc3RydWN0IHsKCWNmZ0ZpbGUgc3RyaW5nCglsb2dnZXIgICpsb2dydXMuRW50cnkKCW5zIE5vZGVTb3VyY2UKCWR0IHRpbWUuRHVyYXRpb24KCXByb2dyZXNzIGJvb2wKfQoKLy8gSW5pdGlhbGl6YXRpb25PcHRpb24gaXMgYW4gb3B0aW9uYWwgc2V0dGluZyB1c2VkIHRvIGluaXRpYWxpemUgdGhlIGNsaWVudAp0eXBlIEluaXRpYWxpemF0aW9uT3B0aW9uIGZ1bmMob3B0cyAqaW5pdE9wdGlvbnMpCgovLyBDb25maWdGaWxlIHNldHMgdGhlIGNvbmZpZyBmaWxlIHRvIHVzZSwgd2hlbiBub3Qgc2V0IHdpbGwgdXNlIHRoZSB1c2VyIGRlZmF1bHQKZnVuYyBDb25maWdGaWxlKGYgc3RyaW5nKSBJbml0aWFsaXphdGlvbk9wdGlvbiB7CglyZXR1cm4gZnVuYyhvICppbml0T3B0aW9ucykgewoJCW8uY2ZnRmlsZSA9IGYKCX0KfQoKLy8gTG9nZ2VyIHNldHMgdGhlIGxvZ2dlciB0byB1c2UgZWxzZSBvbmUgaXMgbWFkZSB2aWEgdGhlIENob3JpYSBmcmFtZXdvcmsKZnVuYyBMb2dnZXIobCAqbG9ncnVzLkVudHJ5KSBJbml0aWFsaXphdGlvbk9wdGlvbiB7CglyZXR1cm4gZnVuYyhvICppbml0T3B0aW9ucykgewoJCW8ubG9nZ2VyID0gbAoJfQp9CgovLyBEaXNjb3Zlcnkgc2V0cyB0aGUgTm9kZVNvdXJjZSB0byB1c2Ugd2hlbiBmaW5kaW5nIG5vZGVzIHRvIG1hbmFnZQpmdW5jIERpc2NvdmVyeShucyBOb2RlU291cmNlKSBJbml0aWFsaXphdGlvbk9wdGlvbiB7CglyZXR1cm4gZnVuYyhvICppbml0T3B0aW9ucykgewoJCW8ubnMgPSBucwoJfQp9CgovLyBQcm9ncmVzcyBlbmFibGVzIGRpc3BsYXlpbmcgYSBwcm9ncmVzcyBiYXIKZnVuYyBQcm9ncmVzcygpIEluaXRpYWxpemF0aW9uT3B0aW9uIHsKCXJldHVybiBmdW5jKG8gKmluaXRPcHRpb25zKSB7CgkJby5wcm9ncmVzcyA9IHRydWUKCX0KfQoKLy8gRGlzY292ZXJ5TWV0aG9kIGFjY2VwdHMgYSBkaXNjb3ZlcnkgbWV0aG9kIG5hbWUgYXMgc3VwcGxpZWQgZnJvbSB0aGUgQ0xJIGFuZCBjb25maWd1cmVzIHRoZSBjb3JyZWN0IE5vZGVTb3VyY2UKLy8gcmV2ZXJ0cyB0byBicm9hZGNhc3QgbWV0aG9kIGlmIGFuIHVuc3VwcG9ydGVkIG1ldGhvZCBpcyBzdXBwbGllZCwgY3VzdG9tIG5vZGUgc291cmNlcyBjYW4gYmUgc2V0IHVzaW5nIERpc2NvdmVyeSgpCmZ1bmMgRGlzY292ZXJ5TWV0aG9kKG0gc3RyaW5nKSBJbml0aWFsaXphdGlvbk9wdGlvbiB7CglyZXR1cm4gZnVuYyhvICppbml0T3B0aW9ucykgewoJCXN3aXRjaCBtIHsKCQljYXNlICJjaG9yaWEiLCAicHVwcGV0ZGIiLCAicGRiIjoKCQkJby5ucyA9ICZQdXBwZXREQk5Te30KCQlkZWZhdWx0OgoJCQlvLm5zID0gJkJyb2FkY2FzdE5Te30KCQl9Cgl9Cn0KCgovLyBEaXNjb3ZlcnlUaW1lb3V0IHNldHMgYSB0aW1lb3V0IGZvciBkaXNjb3ZlcnkgZm9yIHRob3NlIG1ldGhvZHMgdGhhdCBzdXBwb3J0IGl0CmZ1bmMgRGlzY292ZXJ5VGltZW91dCh0IHRpbWUuRHVyYXRpb24pIEluaXRpYWxpemF0aW9uT3B0aW9uIHsKCXJldHVybiBmdW5jKG8gKmluaXRPcHRpb25zKSB7CgkJby5kdCA9IHQKCX0KfQo=",
 	"logging":       "Ly8gZ2VuZXJhdGVkIGNvZGU7IERPIE5PVCBFRElUCgpwYWNrYWdlIHt7IC5QYWNrYWdlIH19CgpmdW5jIChjICp7eyAuRERMLk1ldGFkYXRhLk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1DbGllbnQpIGluZm9mKG1zZyBzdHJpbmcsIGEgLi4uaW50ZXJmYWNle30pIHsKCWMuY2xpZW50T3B0cy5sb2dnZXIuSW5mb2YobXNnLCBhLi4uKQp9CgpmdW5jIChjICp7eyAuRERMLk1ldGFkYXRhLk5hbWUgfCBTbmFrZVRvQ2FtZWwgfX1DbGllbnQpIGVycm9yZihtc2cgc3RyaW5nLCBhIC4uLmludGVyZmFjZXt9KSB7CgljLmNsaWVudE9wdHMubG9nZ2VyLkVycm9yZihtc2csIGEuLi4pCn0K",
