+++ providers/discovery/external/testdata/good.rb
