+++ providers/discovery/external/testdata/good_with_argument.rb
