+++ providers/discovery/flatfile/testdata/badnodes.txt
