+++ providers/discovery/flatfile/testdata/nodes.txt
