+++ providers/discovery/flatfile/testdata/nodes.yaml
