+++ providers/discovery/inventory/choria_mock_test.go
