+++ providers/discovery/inventory/datafile.go
