+++ providers/discovery/inventory/options.go
