+++ providers/discovery/inventory/schema.go
