+++ providers/discovery/inventory/testdata/good-inventory.yaml
