+++ providers/discovery/puppetdb/puppetdb.go
